Topic3Tools/pyproject.toml
[project]
name = "topic3tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.1",
    "datasets>=4.5.0",
    "fortune>=1.1.1",
    "huggingface-hub>=1.3.7",
    "langchain>=1.2.8",
    "langchain-openai>=1.1.7",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "matplotlib>=3.10.8",
    "numexpr>=2.14.1",
    "numpy>=2.2.6",
    "openai>=2.16.0",
    "requests>=2.32.5",
    "tenacity>=9.1.2",
    "torch>=2.10.0",
    "tqdm>=4.67.2",
    "transformers>=5.0.0",
]
